add avgspeed to barcharts and heightgraph
As it would be difficult to find colors for every returned speed, the avgspeed values
are grouped in categories (see getSpeedRange function in ors-utils-service.js and mappings.js.
Adjusted ors-route-extras.js and ors-route-service so the heightgraph and barchart will take the categories.

Fixed a typo in ors-route-service that prevented red steepness values from showing.

closes https://github.com/GIScience/openrouteservice/issues/139