rework geocode with pelias endpoints
geocoding requests now are sent to host/geocode/search? and host/geocode/reverse? enpoints
instead of host/geocode.

Functionality is the same as before.
Additionally the center of the current map view is used as a focus point by default.
Objects within 100km of this point are more relevant. Importance decays from the center outward.
Overall accuracy should be enhanced with this.

closes https://github.com/GIScience/openrouteservice-app/issues/196