ors-header.js: Rework ctrl.changeOptions
ctrl.changeOptions formerly only detected language change. It was reworked to detect changes other settings as well.
It will now also broadcast the changed settings to the rootScope for the changes to take immediate effect.
The call of ctrl.changeOptions was modified in settings.html for that purpose.

refers to https://github.com/GIScience/openrouteservice-app/issues/234