Set heightgraph visibility & distance marker settings otf
it is now possible to set the visibility of the heightgraph from the settings.
This as well as the distance marker setting is reflected instantly on the map.

As the Heightgraph setting is stored in the cookies now they are checked
before checking the device screen size.
The heightgraph can be viewed on mobile devices now but the design is
still not responsive.

refers to https://github.com/GIScience/openrouteservice-app/issues/234