Add Developer Option to set custom endpoints
The Developer Settings are accessible through a barely visible gear icon
in the bottom left corner of the settings modal.
There are presets for setting the endpoits to a local ORS backend or to the ORS API.
Changing the baseURL input will adjust the baseURL of every endpoint.
It is also possible to adjust the enpoints individually by clicking 'Specify Endpoints'.
Upon clicking 'Change Endpoints' the requests will be sent to the new endpoints.
'Reset' will switch all endpoints to their initial state.
Happy developing.

Refers to https://github.com/GIScience/openrouteservice-app/issues/243