Fix poi service error
when using the poi service and a feature without
osm tags was returned, the code handeling reading the
features would not check if there actually are osm tags.

This would result in an error in the console and the ui
being stuck in loading of pois.

Fixed by checking for osmTags before accessing them

Fixes https://github.com/GIScience/openrouteservice-app/issues/363